Rewrote training to support sweeps
The `wandb` client exhibits strange behaviour when individual folds are
to be logged during a sweep: essentially, only a single fold is stored.
To remedy this, the `fold` may now be controlled during the sweep, thus
making it possible to have a *single* run.

(additional updates were made in order to provide better grouping)